requirements_whisper.txt
openai-whisper>=20231117
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0
ffmpeg-python>=0.2.0
